+++ src/lib/pages/home/components/index.ts
@@ -1,5 +1,4 @@
 export * from "./CardInfo";
 export * from "./DevShortcut";
-export * from "./TopDecorations";
 export * from "./QuickMenuLite";
 export * from "./QuickMenuMobileLite";